fw-6473, move base tests for uncontrolled site content APIs to own module
