+++ firstvoices/backend/tests/test_apis/base/base_media_test.py
@@ -12,7 +12,9 @@ from backend.tests.factories import (
     get_site_with_authenticated_nonmember,
     get_site_with_staff_user,
 )
-from backend.tests.test_apis.base import base_api_test
+from backend.tests.test_apis.base.base_uncontrolled_site_api import (
+    BaseUncontrolledSiteContentApiTest,
+)
 from backend.tests.utils import get_sample_file
 
 VIMEO_VIDEO_LINK = "https://vimeo.com/226053498"
@@ -480,7 +482,7 @@ class RelatedMediaTestMixin(MediaTestMixin):
 class BaseMediaApiTest(
     MediaTestMixin,
     FormDataMixin,
-    base_api_test.BaseUncontrolledSiteContentApiTest,
+    BaseUncontrolledSiteContentApiTest,
 ):
     """
     Tests for the list, detail, create, and delete APIs for media endpoints.
