+++ firstvoices/backend/tests/test_apis/test_bulkvisibility_api.py
@@ -6,13 +6,15 @@ from backend.models.constants import Visibility
 from backend.models.jobs import BulkVisibilityJob
 from backend.tests import factories
 from backend.tests.test_apis.base.base_api_test import (
-    BaseReadOnlyUncontrolledSiteContentApiTest,
-    SiteContentCreateApiTestMixin,
     SuperAdminAsyncJobPermissionsMixin,
     WriteApiTestMixin,
 )
 
 from ..test_search_indexing.base_indexing_tests import TransactionOnCommitMixin
+from .base.base_uncontrolled_site_api import (
+    BaseReadOnlyUncontrolledSiteContentApiTest,
+    SiteContentCreateApiTestMixin,
+)
 
 
 class TestBulkVisibilityEndpoints(
