+++ firstvoices/backend/tests/test_apis/test_dictionary_cleanup_api.py
@@ -7,10 +7,12 @@ from backend.models.jobs import DictionaryCleanupJob, JobStatus
 from backend.tasks.dictionary_cleanup_tasks import cleanup_dictionary
 from backend.tests import factories
 from backend.tests.test_apis.base.base_api_test import (
-    BaseReadOnlyUncontrolledSiteContentApiTest,
-    SiteContentCreateApiTestMixin,
     SuperAdminAsyncJobPermissionsMixin,
     WriteApiTestMixin,
+)
+from backend.tests.test_apis.base.base_uncontrolled_site_api import (
+    BaseReadOnlyUncontrolledSiteContentApiTest,
+    SiteContentCreateApiTestMixin,
 )
 from backend.tests.test_search_indexing.base_indexing_tests import (
     TransactionOnCommitMixin,
