+++ firstvoices/backend/tests/test_apis/test_ignored_characters_api.py
@@ -1,5 +1,5 @@
 from backend.tests import factories
-from backend.tests.test_apis.base.base_api_test import (
+from backend.tests.test_apis.base.base_uncontrolled_site_api import (
     BaseReadOnlyUncontrolledSiteContentApiTest,
 )
