+++ firstvoices/backend/tests/test_apis/test_join_request_api.py
@@ -9,11 +9,11 @@ from backend.models import JoinRequest, Membership
 from backend.models.constants import AppRole, Role, Visibility
 from backend.models.join_request import JoinRequestReasonChoices, JoinRequestStatus
 from backend.tests import factories
-from backend.tests.test_apis.base.base_api_test import (
+from backend.tests.test_apis.base.base_api_test import WriteApiTestMixin
+from backend.tests.test_apis.base.base_uncontrolled_site_api import (
     BaseReadOnlyUncontrolledSiteContentApiTest,
     SiteContentCreateApiTestMixin,
     SiteContentDestroyApiTestMixin,
-    WriteApiTestMixin,
 )
 from backend.views.utils import get_site_url_from_appjson
