+++ firstvoices/backend/tests/test_apis/test_person_api.py
@@ -1,7 +1,7 @@
 from backend.models.media import Person
 from backend.tests import factories
 from backend.tests.factories import PersonFactory
-from backend.tests.test_apis.base.base_api_test import (
+from backend.tests.test_apis.base.base_uncontrolled_site_api import (
     BaseUncontrolledSiteContentApiTest,
 )
