+++ firstvoices/backend/tests/test_apis/test_site_features_api.py
@@ -3,7 +3,7 @@ import pytest
 from backend.models.constants import AppRole, Visibility
 from backend.models.sites import SiteFeature
 from backend.tests import factories
-from backend.tests.test_apis.base.base_api_test import (
+from backend.tests.test_apis.base.base_uncontrolled_site_api import (
     BaseUncontrolledSiteContentApiTest,
 )
