+++ firstvoices/backend/tests/test_apis/test_stats_api.py
@@ -5,7 +5,7 @@ import pytest
 from backend.models.constants import Role, Visibility
 from backend.models.dictionary import TypeOfDictionaryEntry
 from backend.tests import factories
-from backend.tests.test_apis.base.base_api_test import (
+from backend.tests.test_apis.base.base_uncontrolled_site_api import (
     BaseSiteContentApiTest,
     SiteContentListApiTestMixin,
 )
